% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{NCREIF}
\alias{NCREIF}
\title{Main NCREIF Query Function}
\usage{
NCREIF(SelectString, WhereString, GroupByString, Username, Password, verbose)
}
\arguments{
\item{SelectString}{SQl Select String for API, does not include select keyword. Fields are Bracketed. All Fields must be in aggregate function.}

\item{WhereString}{SQL string for where parameters of query. All Fields must be bracketed.}

\item{GroupByString}{SQL Group by fields. Each field must be bracketed, seperated with commas. Must include at least one group by field.}

\item{Username}{NCREIF username}

\item{Password}{NCREIF Password}

\item{Verbose}{True/False for development.}
}
\description{
This Function will query the NCREIF SOAP
}
\examples{
SelectString="SUM([NOI]) AS NOI,SUM([CapEx])AS CapEx, SUM([MV]) AS MV,SUM([MVLag1])AS MVLag1, SUM([PSales])AS PSales, Round(SUM([Denom]),4) AS Denom, Round(Sum([NOI]) / Sum([Denom]),4) AS IncomeReturn,Round((Sum([MV])-Sum([MVLag1])-Sum([CapEx])+Sum([PSALES]))/(Sum([DENOM])),4) AS CapitalReturn, Round((Sum([NOI]) / Sum([Denom]))+(Sum([MV])-Sum([MVLag1])-Sum([CapEx])+Sum([PSALES]))/(Sum([DENOM])),4) AS TotalReturn,Count([MV]) AS PropCount"
WhereString="[NPI]=1"
GroupByString="[Period],[YYYYQ] ,[Year],[Quarter]"
NCREIF(SelectString,WhereString,GroupByString,readline("Enter Username"),readline("Enter Password"),FALSE)
}
